Serve OpenAPI files + update api types
