+++ go.mod
@@ -3,23 +3,24 @@ module EtuEDT-Go
 go 1.21
 
 require (
-	github.com/arran4/golang-ical v0.1.0
-	github.com/go-co-op/gocron v1.33.1
-	github.com/gofiber/fiber/v2 v2.50.0
+	github.com/arran4/golang-ical v0.2.2
+	github.com/go-co-op/gocron v1.36.1
+	github.com/gofiber/fiber/v2 v2.51.0
 )
 
 require (
-	github.com/andybalholm/brotli v1.0.5 // indirect
-	github.com/google/uuid v1.3.1 // indirect
-	github.com/klauspost/compress v1.16.7 // indirect
+	github.com/andybalholm/brotli v1.0.6 // indirect
+	github.com/google/uuid v1.4.0 // indirect
+	github.com/klauspost/compress v1.17.4 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.19 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mattn/go-runewidth v0.0.15 // indirect
-	github.com/rivo/uniseg v0.2.0 // indirect
+	github.com/rivo/uniseg v0.4.4 // indirect
 	github.com/robfig/cron/v3 v3.0.1 // indirect
+	github.com/stretchr/testify v1.8.4 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
-	github.com/valyala/fasthttp v1.50.0 // indirect
+	github.com/valyala/fasthttp v1.51.0 // indirect
 	github.com/valyala/tcplisten v1.0.0 // indirect
-	go.uber.org/atomic v1.9.0 // indirect
-	golang.org/x/sys v0.13.0 // indirect
+	go.uber.org/atomic v1.11.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
 )
